// src/enums/ClothingType.ts
export enum ClothingType {
    BUTTON_DOWN = 'Button down',
    CROP_TOP = 'Crop top',
    DRESS = 'Dress',
    PANTS = 'Pants',
    JEWELRY = 'Jewelry',
    LEGGINGS = 'Leggings',
    LONG_SLEEVE = 'Long sleeve',
    MEN_FORMAL_WEAR = 'Men formal wear',
    PARTY_CLOTHES = 'Party clothes',
    SHOES = 'Shoes',
    SHORTS = 'Shorts',
    SKIRT = 'Skirt',
    SWEATER = 'Sweater',
    SWEAT_PANTS = 'Sweat pants',
    SWEATSHIRT = 'Sweatshirt',
    TANK_TOP = 'Tank top',
    TSHIRT = 'Tshirt',
    HAT = 'Hat',
    JACKET = 'Jacket',
}